Script for GFP in mid of deg in mouse.
mRNA extraction were collected from sorting cells in flow cytometry. This could be a reason why the quality of mRNA is not quite good. In addition, in each group, the samples come from independent mouse. Thus, it is hard to group samples (more details could be seen in PCA plot).